some more restructiring
