+++ .gitignore
@@ -1,2 +1,2 @@
 *.class
-.idea/workspace.xml
+.idea/*.xml
