+++ Main.java
@@ -3,18 +3,21 @@
 public class Main{
 
 	public static void main(String[] args) {
+		final double DEFAULT_MULTPLIER = 1;
+		final double UPPER_BOUND = Note.DEFAULT_FREQ * Note.DEFAULT_FREQ;
+
 
 		MusicSheet sheet = new MusicSheet();
 		double f = Note.DEFAULT_FREQ;
-		for(int i = 0; i < 100 ; i++){
-			f += NoteGenerator.quadraticNote(i);
+		for(int i = 0; i < 100  && f < UPPER_BOUND; i++){
+			f += NoteGenerator.getNoteOfFunction('2', DEFAULT_MULTPLIER, i);
 			sheet.addNote(f, Note.DEFAULT_DURATION);
 		}
 
-		StdAudio.save("graph_samples/quadratic.wav", sheet.getArray());
+		StdFileIO.save("graph_samples/quadratic.wav", sheet.getArray());
 
 		System.out.println("finished.");
-
+		StdAudio.close();
 	}
 
 }
